Refactor createPortalClient for node building from memory
