+++ packages/portal-client/src/services/portalNetwork/client.ts
@@ -9,27 +9,31 @@ import {
 } from 'portalnetwork'
 import { createDatabase } from './db'
 import { TauriUDPTransportService } from './transports'
+import { DEFAULT_DB_SIZE, STARTUP_DELAY_MS } from '@/utils/constants/config'
 
 import type { PortalNetwork } from 'portalnetwork'
 import type { Multiaddr } from '@multiformats/multiaddr'
-// const isBrowser = () => !window.__TAURI__
 
-const db = createDatabase('network_db', { prefix: 'portalclient_' })
+let client: PortalNetwork
 
 export const createPortalClient = async (port: number): Promise<PortalNetwork> => {
-  try {
+  const db = createDatabase('network_db', { prefix: 'portalclient_' })
+  
+  const createNetwork = async (rebuildFromMemory: boolean): Promise<PortalNetwork> => {
     const privateKey = await keys.generateKeyPair('secp256k1')
     const enr = SignableENR.createFromPrivateKey(privateKey)
     const nodeAddr = multiaddr(`/ip4/0.0.0.0/udp/${port}`)
     enr.setLocationMultiaddr(nodeAddr)
-    const client = await createPortalNetwork({
+    
+    return createPortalNetwork({
       transport: TransportLayer.TAURI,
       supportedNetworks: [
         { networkId: NetworkId.HistoryNetwork },
         { networkId: NetworkId.StateNetwork },
       ],
       db,
-      dbSize: async () => 1024 * 1024 * 1024,
+      rebuildFromMemory,
+      dbSize: async () => DEFAULT_DB_SIZE,
       config: {
         enr,
         bindAddrs: { ip4: nodeAddr },
@@ -42,14 +46,19 @@ export const createPortalClient = async (port: number): Promise<PortalNetwork> =
         },
       },
     })
-
-    await client.start()
-    await client.bootstrap()
-    
-    await new Promise(resolve => setTimeout(resolve, 1000))
-    
-    return client
-  } catch (error) {
-    throw error
   }
+  
+  try {
+    client = await createNetwork(true)
+  } catch (error) {
+    console.error('Failed to create node from memory, creating from scratch:', error)
+    client = await createNetwork(false)
+  }
+  
+  await client.start()
+  await client.storeNodeDetails()
+  
+  await new Promise(resolve => setTimeout(resolve, STARTUP_DELAY_MS))
+  
+  return client
 }
