+++ packages/portal-client/src/services/portalNetwork/types.ts
@@ -1,7 +1,8 @@
-export const APPROVED_METHODS = [
-  'eth_getBlockByHash',
-  'eth_getBlockByNumber',
-  // 'portal_findNodes',
-  // 'eth_getBlockReceipts',
-  // 'eth_getLogs',
-] as const
+export interface MethodConfig {
+  name: string
+  paramPlaceholder: string
+  handler: (
+    input: string, 
+    sendRequest: (method: string, params?: any[])
+    => Promise<any>) => void | Promise<any>
+}
