+++ packages/portal-client/src/utils/constants/config.ts
@@ -1,5 +1,8 @@
 import { ConfigId } from '../types'
 
+export const DEFAULT_DB_SIZE = 1024 * 1024 * 1024
+export const STARTUP_DELAY_MS = 1000
+
 export const CONFIG_DEFAULTS: { id: ConfigId; title: string; defaultValue: string; description: string }[] = [
   {
     id: ConfigId.UdpPort,
